splice: return type ssize_t from all helpers
Mainline: 0f292086c22b43202daffc14b585d3b54b9a1206
From: v6.8-rc1
Severity: Low

task: 522091

Not sure why some splice helpers return long, maybe historic reasons.
Change them all to return ssize_t to conform to the splice methods and
to the rest of the helpers.

Suggested-by: Christian Brauner <brauner@kernel.org>
Link: https://lore.kernel.org/r/20231208-horchen-helium-d3ec1535ede5@brauner/
Link: https://lore.kernel.org/r/20231212094440.250945-2-amir73il@gmail.com

(backported from commit 0f292086c22b43202daffc14b585d3b54b9a1206)

Reviewed-on: https://code.kylinos.cn:443/c/institute/kernel/klinux/+/112233
Tested-by: kernel-bot <kernel-bot@kylinos.cn>
Code-Style: kernel-bot <kernel-bot@kylinos.cn>

+++ fs/internal.h
@@ -247,10 +247,10 @@ int do_statx_fd(int fd, unsigned int flags, unsigned int mask,
 /*
  * fs/splice.c:
  */
-long splice_file_to_pipe(struct file *in,
-			 struct pipe_inode_info *opipe,
-			 loff_t *offset,
-			 size_t len, unsigned int flags);
+ssize_t splice_file_to_pipe(struct file *in,
+			    struct pipe_inode_info *opipe,
+			    loff_t *offset,
+			    size_t len, unsigned int flags);
 
 /*
  * fs/xattr.c:
